+++ spec/features/changement_subscription_spec.rb
@@ -1,12 +1,16 @@
 # frozen_string_literal: true
 
-RSpec.describe 'Dependency observers (#observe, #unobserve, #clear_listeners)' do
+RSpec.describe 'Changement subscription (#observe, #unobserve, #clear_observers)' do
   specify 'subscribe (#observe) to dependency changement (namespaces and deps) and listen' do
     container = SmartCore::Container.define do
       namespace(:storages)  do
         register(:database) { 'database' }
         register(:cache) { 'cache' }
         register(:replica) { 'replica' }
+
+        namespace(:creds) do
+          register(:redis) { 'redis' }
+        end
       end
     end
 
@@ -16,38 +20,44 @@ RSpec.describe 'Dependency observers (#observe, #unobserve, #clear_listeners)' d
 
     container.observe('storages.database') { database_interceptor << 'db_changed!' }
     container.observe('storages.cache') { cache_interceptor << 'cache_changed!' }
-    container.observe('storages') { namespace_interceptor << 'namespace_changed!' }
+    container.observe('storages.creds') { namespace_interceptor << 'namespace_changed!' }
 
     expect(database_interceptor).to be_empty
     expect(cache_interceptor).to be_empty
     expect(namespace_interceptor).to be_empty
 
-    container.register('storages.database') { 'new_database' }
+    container.fetch(:storages).register(:database) { 'new_database' }
+
     expect(database_interceptor).to contain_exactly('db_changed!')
     expect(cache_interceptor).to be_empty
     expect(namespace_interceptor).to be_empty
 
-    container.register('storages.cache') { 'new_cache' }
+    container.fetch('storages').register(:cache) { 'new_cache' }
     expect(database_interceptor).to contain_exactly('db_changed!')
     expect(cache_interceptor).to contain_exactly('cache_changed!')
     expect(namespace_interceptor).to be_empty
 
-    container.register('storages.database') { 'new_database' }
+    container.fetch(:storages).register(:database) { 'new_database' }
     expect(database_interceptor).to contain_exactly('db_changed!', 'db_changed!')
     expect(cache_interceptor).to contain_exactly('cache_changed!')
     expect(namespace_interceptor).to be_empty
 
-    container.register('storages.cache') { 'new_cache' }
+    container.fetch(:storages).register(:cache) { 'new_cache' }
     expect(database_interceptor).to contain_exactly('db_changed!', 'db_changed!')
     expect(cache_interceptor).to contain_exactly('cache_changed!', 'cache_changed!')
     expect(namespace_interceptor).to be_empty
 
-    container.register('storages.replica') { 'new_replica' }
+    container.fetch(:storages).register(:replica) { 'new_replica' }
     expect(database_interceptor).to contain_exactly('db_changed!', 'db_changed!')
     expect(cache_interceptor).to contain_exactly('cache_changed!', 'cache_changed!')
     expect(namespace_interceptor).to be_empty
 
-    container.namespace('storages') {}
+    container.fetch(:storages).namespace(:creds) {}
+    expect(database_interceptor).to contain_exactly('db_changed!', 'db_changed!')
+    expect(cache_interceptor).to contain_exactly('cache_changed!', 'cache_changed!')
+    expect(namespace_interceptor).to contain_exactly('namespace_changed!')
+
+    container.namespace(:storages) {}
     expect(database_interceptor).to contain_exactly('db_changed!', 'db_changed!')
     expect(cache_interceptor).to contain_exactly('cache_changed!', 'cache_changed!')
     expect(namespace_interceptor).to contain_exactly('namespace_changed!')
@@ -79,8 +89,8 @@ RSpec.describe 'Dependency observers (#observe, #unobserve, #clear_listeners)' d
     # rubocop:enable Lint/UselessAssignment
 
     # register new entities
-    container.register('api.google') { 'new_google' }
-    container.register('api.kickbox') { 'new_kickbox' }
+    container.fetch(:api).register('google') { 'new_google' }
+    container.fetch(:api).register('kickbox') { 'new_kickbox' }
 
     # received dependency entity change
     expect(google_interceptor).to contain_exactly("api.google__#{container.object_id}")
@@ -92,8 +102,8 @@ RSpec.describe 'Dependency observers (#observe, #unobserve, #clear_listeners)' d
     expect(container.unobserve(google_observer)).to eq(true) # unsubscribed (true)
     expect(container.unobserve(google_observer)).to eq(false) # already unsubscribed (false)
 
-    container.register('api.google') { 'another_new_google' }
-    container.register('api.kickbox') { 'another_new_kickbox' }
+    container.fetch('api').register(:google) { 'another_new_google' }
+    container.fetch('api').register(:kickbox) { 'another_new_kickbox' }
 
     # unsubscribed
     expect(google_interceptor).to contain_exactly("api.google__#{container.object_id}")
@@ -118,8 +128,8 @@ RSpec.describe 'Dependency observers (#observe, #unobserve, #clear_listeners)' d
     expect(namespace_interceptor).to contain_exactly("api__#{container.object_id}")
 
     # register new entities in new enty namespace
-    container.register('api.google') { 'another_another_new_google' }
-    container.register('api.kickbox') { 'another_another_new_kickbox' }
+    container.fetch(:api).register(:google) { 'another_another_new_google' }
+    container.fetch(:api).register(:kickbox) { 'another_another_new_kickbox' }
 
     # unsubscribed
     expect(google_interceptor).to contain_exactly("api.google__#{container.object_id}")
@@ -135,8 +145,8 @@ RSpec.describe 'Dependency observers (#observe, #unobserve, #clear_listeners)' d
     # unsubscribe all kickbox observers
     container.clear_observers('api.kickbox')
     # register new entities
-    container.register('api.google') { 'another_another_new_google' }
-    container.register('api.kickbox') { 'another_another_new_kickbox' }
+    container.fetch(:api).register(:google) { 'another_another_new_google' }
+    container.fetch(:api).register(:kickbox) { 'another_another_new_kickbox' }
 
     # unsubscribed
     expect(google_interceptor).to contain_exactly("api.google__#{container.object_id}")
@@ -152,8 +162,8 @@ RSpec.describe 'Dependency observers (#observe, #unobserve, #clear_listeners)' d
     # unsubscribe all
     container.clear_observers
 
-    container.register('api.google') { 'another_new_google' }
-    container.register('api.kickbox') { 'another_new_kickbox' }
+    container.fetch(:api).register(:google) { 'another_new_google' }
+    container.fetch(:api).register(:kickbox) { 'another_new_kickbox' }
     container.namespace('api') {}
 
     # unsubscribed